Removed child functionality and added ronin art.
Ultimately I don't have any non-moving or nested childs in the demo I'm making, so I decided to remove it.

This also has to do with the fact that I've been having a hard time finding a good implementation for the child system.

I'm planning to replace this functionality simply with function calls like ToLocal(transform, pos).