Implemented observable pattern for collisions.
I've thought about what is best when discussing collisions, but I feel like the observer pattern can work for ecs. 

In my mind, not nessecary the systems will be subscirbers (as they have lazy initialization this can cause weird issues, so I'd rather avoid it), but dedicated classes will be made which will handle subscription behaviour at the end of the frame.